requirements.txt
requests
python-dotenv
ebaysdk
